refactor(ui): 发射器移入 ui/sender/ 子包，组件放入 components/
