+++ src/danmaku_sender/ui/sender/__init__.py
@@ -1,0 +1,3 @@
+from .page import SenderPage
+
+__all__ = ["SenderPage"]
