+++ src/danmaku_sender/ui/sender/components/strategy_tabs.py
@@ -3,9 +3,8 @@ from PySide6.QtWidgets import (
     QSpinBox, QDoubleSpinBox, QFrame
 )
 
-from .framework.binder import UIBinder
-
 from danmaku_sender.runtime.state.app_state import AppState
+from danmaku_sender.ui.framework.binder import UIBinder
 
 
 class StrategySettingsTabs(QTabWidget):
